Fixed bug with SHOP causing other conds to break
The 'inStore" flag here has absolutely nothing to do with the NPC shops. It's an attribute that gets set when an item first drops from a mob, chest, etc. So this was pretty much _always_ returning true.

There's no reason to have a packet check for this condition anyway so we can just always return false

Also added a few comments to make these attributes a bit more clear