Add loot filter output builder
Currently, all filter output rules are parsed as an item requires a
name or description. This adds a builder to parse all output rules in
advance. Every keyword, static or dynamic, matches a simple regex. The
matches from this regex are used to create functions which yield a
string based on the match: name, param1, param2. These functions share a
common context for each rule to allow for stateful changes. When a rule is applied
the return values are concatenated to obtain the rule's output.

Included in this patch is a simple comparison between the new and old
results. A breakpoint may be set to observe erroneous strings. The new
name or description is replaced with an error string too.