Changes to prevent exceptions when using an invalid soundID
Default soundID on an action will be a valid soundID instead of -1 incase the soundID is used.

Parsed soundID will be 0 if the soundID exceeds the MAX_SOUND_ID index. This value(MAX_SOUND_ID) unfortunately will need to be updated if any sounds are added or removed from the Sounds.txt.